requirements.txt
GitPython==0.3.5
click==3.3